Consolidate generating random schedule entry for tests into test helper function.
